requirements.txt
asgiref==3.4.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
coverage==6.1.1
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.9
django-allauth==0.45.0
django-money==2.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==9.7.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.9.3
oauthlib==3.1.1
packaging==21.2
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
py-moneyed==1.2
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
